+++ Changelog.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file!
 
 ### Changed
 
-- Changed: No notable changes
+- Changed: Code optimized for performance.
 
 ### Fixed